Fix load from commandline
during song load, reset_ports was called and this
resulted in the jack port being reset to 0 in
the m_output_ports map.

This was causing jack port to be not found
in Sequencer::sequence_midi and this no
midi events were generated.

This changeset removes Sequencer::reset_ports()
because it seems to be not needed. The signal
for track add is properly connected and does its job.

This changeset also adds a check in Sequencer::track_added()
that prevents the jack port handle leak to happen again.